requirements.txt
numpy
python-vlc

